Don't generate bindings for standard Wasm C API
Unfortunately this now means that we require nightly Rust to build the
C API because of the use of an unstable feature when expanding.  This
is required as this is the only way to apply parse-specific features.

Alterantively, we can split the wasm-c-api out as a separate crate and
include it as a dependency, this should allow this to continue to work
on stable Rust.