Cache lookup of VMMemoryDefinition.
To add to the cache, mutability is required, so now `resolve_memory_ptr` must be `&mut self` which means that `translate_operator` may not hold those convenient immutable references across the whole function.

The search and replace to implement the above changes affected some commented out code, that code is removed.

MemoryStyle grows PartialEq and Eq so that we can check what the memory style is without a match expression.