feat(wasm-common) Support `Result` as returned type for host functions.
We re-use the `TrapEarly` trait to represent a host function that
return a Wasm value, or a `Result` of a Wasm value.

In this patch, the returned value of the `wrap` functions is
`Result<Rets::CStruct, Trap::Error>`. Note that it will require a
large amount of modification on the compilers side.