feat(api) Reimplement `WasmExternType` as `FromToNativeWasmType`.
In addition to a more self-explanatory renaming from `WasmExternType`
to `FromToNativeWasmType`, this patch updates the `to_native` and the
`from_native` methods to panic when the values cannot be converted
from one type to another by using their respective `TryInto`
implementations.

For instance, `7i8.to_native()` equals to `7i32`, but
`u32::MAX.to_native()` cannot be converted to `i32`, and so it will
panic with the message:

> cannot convert 4294967295u32 to i32.

The old implementation was returning `-1i32`, which was incorrect.