fix: `UnwindRegistry` owns the `eh_frame`s.
`UnwindRegistry` registers `eh_frame` from a slice. This slice is a
pointer owned by `SerializableModule`, which is held by `Artifact`. If
the artifact is dropped _before_ the `UnwindRegistry`, then we have a
dangling pointer, and it can crash. Actually, it crashes, see
https://github.com/wasmerio/wasmer/pull/1581 as our first attempt to
fix this problem.

So instead of `UnwindRegistry` to register a pointer to a data it
doesn't own, this patch updates `UnwindRegistry` to register a pointer
to data that it owns. `UnwindRegistry` now receives an `eh_frame` of
kind `Vec<u8>` instead of `&[u8]`. The bytes are copied.

In addition, this patch creates an `UnwindRegistryExt` trait, so that
it brings consistency amongst all `UnwindRegistry` implementation
(`systemv`, `windows_x64` and `dummy`). This is for an internal usage
only.