Ignore mapdir_with_leading_slash on Windows for now
These tests are failing because we call `take(N)` on chars of a string
read from a file. It seems that on Windows, the programs get line
endings with `\r\n`, this is a problem for the way our tests assert
correct output.

We'll need to find a general solution to this problem later.