test: Remove a warning fix for Windows.
Windows will raise a warning because `clang++` is used to handle C
code, so I previously add `-x c` to tell `clang++` to handle code as C
code. Bbut if code is treated as C, it will fail because
`static_assert` is not available for C on Windows.

So let's live with the warning.