feat(wasi) Allow the `=` sign in the environment variable value.
Having `a=b` means that the environment variable key is `a`, and its
value is `b`.

Having `a=b=c` means that the key is `a` and its value is `b=c`.

It's OK to get this, e.g. within a CGI context where values can hold
strings such as `sec-ch-ua: "Chromium;v=86"` etc. See
https://github.com/wasmerio/wasmer/issues/1708 for a longer
explanation.