feat(wasi) Seperate env's key and value and do different validations.
This patch keeps the env's key and value separate until the `build()`
method is called. It allows to apply different validations resp. for
the key and the value: Both must not contain a nul byte, but only the
key cannot contain an equal sign.

This patch also updates the test cases accordingly.

Finally this patch updates the `envs` and `args` methods to reuse
respectively `env` and `arg` to avoid code repetition.