fix(c-api) `wasm_importtype_t` _owns_ its fields.
The `wasm.h` file defines `wasm_importtype_new` to take ownership of
the module, the name, and the extern type. This patch thus changes
`NonNull` to `Box` for those types.

`wasm_importtype_new` also handles null pointer with `Option<Box<T>>`
and returns an `Option<Box<T>>`.

The rest of the code is simplified. The patch fixes also some typos in
variable namings.