feat(vm) Remove `InstanceHandle::from_vmctx`.
The `InstanceHandle::from_vmctx` method is only used by
`CallThreadState::any_instance`, where a given closure expect an
`&InstanceHandle`. However, the only closure that exists in the code
only reads an `&Instance` from the `InstanceHandle`.

So… this patch updates this only closure and `any_instance` to work on
an `&Instance` directly, which allows us to remove
`InstanceHandle::from_vmctx`. It results to less code and less pointer
arithmetic.