feat(vm) Change how the layout of `Instance` is calculated.
The old code is working because the alignment of `Instance` is 16, and
its size is 400. Even with a really large `VMContext`, it will still
fit in an alignment of 8. But the way it is calculated is wrong. The
size is correctly calculated, but the alignment misses `VMContext`.

Quoting our own documentation:

> It is important to remind that Instance is dynamically-sized based
> on `VMOffsets`: The `Instance.vmctx` field represents a
> dynamically-sized array that extends beyond the nominal end of the
> type.

So first, let's compute the layout of the `Instance.vmctx` “array”
with `Layout::array`. Second, let's compute the layout of `Instance`
itself with `Layout::new`. And third, let's compute the entire layout
by using `Layout::extend` + `Layout::pad_to_align`.