Continue fixing linking errors that come up when building hello.cc.
This patch includes a smattering of fixes. Further the implementation of copy(),
return Wasmer types instead of the base types, add missing destroy(), add Limits
to WasmerMemoryType so that we can return a reference to it, expose the
WasmerRefWrapper constructor through to WasmerExternWrapper, flesh out
WasmerModule and WasmerInstance some more, add an assert to WasmerExternWrapper.