fix(c-api) Fix how `wasm_functype_t` is implemented.
In `wasm.h`, `wasm_functype_t` is implemented with
`WASM_DECLARE_TYPE`, so with `WASM_DECLARE_VEC(functype, *)`. This `*`
means the C struct for the vector is defined:

```c
struct wasm_functype_vec_t {
    size_t size;
    wasm_functype_t** data;
}
```

The way we implement `wasm_functype_vec_t` in Rust is with the
`wasm_declare_vec!` macro. And it is wrong. We must use
`wasm_declared_boxed_vec!`.