feat(tracing) Introducing `wasmer-tracing`.
This is a proof-of-concept, or at best a work-in-progress, of
integrating USDT probes inside Wasmer through the new `wasmer-tracing`
crate.

USDT stands for Userland Statically Defined Tracing. USDT probes are
probes originally designed by DTrace, but now sucessfully implemented
in many environments, like Darwin, Solaris, Linux, Illumos, and even
[Windows][windows-dtrace]. It allows dynamic tracing at low-cost (a
probe is a noop, except when run with `dtrace` where a special section
of the object file is read to get the real probes and to instrument
the executable code, in some cases).

USDT aren't tied to DTrace only now. On Linux for example, eBPF
understands USDT probes. The process of generating and linking the
probes is automated, but the FFI bridge (`probes_ffi.c` and `lib.rs`)
must be completed by hand for each new added probe. It's not an issue
as a few number of strategic probes are going to be implemented.

This patch provides a new crate: `wasmer-tracing`.

The current set of probes (`instance_start`, `instance_end`,
`function_start`, `function_invoked2` and `function_end`) is only
present to showcase the idea.

The `scripts/tracing.d` script showcases how to read data from those
probes to trace the execution of the VM itself with a DTrace script.

Note that for the moment, the probes —as implemented— are designed to
be used in Rust. It means we can only analyze the
compilation-step/compile-time of a Wasm module into executable code,
and some runtime operations. More work is required to implement probes
inside the executable code we generate with the compilers, i.e for the
entire execution-step/runtime (LLVM has [XRay Instrumentation] for
example, not sure it's the path we must follow but it exists).

The next commit showcases how to use `wasmer-tracing`.

[windows-dtrace]: https://techcommunity.microsoft.com/t5/windows-kernel-internals/dtrace-on-windows/ba-p/362902
[XRay Instrumentation]: https://www.llvm.org/docs/XRay.html