chore(vm) Rewrite to respect C coding style.
This patch renames `RegisterSetjmp` and `Unwind` to `register_setjmp`
and `unwind`. It also renames some variables. The goal is to respect
the C coding style.

The `trapcode` and `traphandlers` modules are renamed `code` and
`handlers` respectively. The symbol was exported by the `trap` module,
so nothing changes from the public scope.